[CB-4976] Don't symlink into ~/.cordova/lib for local libs
If the user specifies a local URI in his config.json, we don't want to
symlink that lib into $HOME/.cordova/lib, just use it in-place as the
source of the platform.

lazy_load calls now return a promise for the path to the target, and all
logic in other modules that was recomputing the path into the global lib
directory has been fixed to just use that return value. This is more
flexible and supports this feature nicely.