ios: add support for TargetDevice preference
Allows the user to configure which category of devices they are
targeting (handset, tablet, or universal). Without this setting, a user
would not be able to use the CLI workflow to develop an app for
publishing in the iOS app store without supporting all of the iOS
devices. This allows a user to target just iPhones/iPods or iPads if
they so desire. The app store has requirements that you properly support
the platforms you target and thus it may be prudent to limit your
supported devices.

<preference name="TargetDevice" value="handset"/>

Valid values are handset, tablet, or universal.