CB-5282 [CLI] deal with env variable, and fixes
Based on feedback, and some corrections by my own.
Fixed:
- No more tabs.
- Brand misspelled
- Included new line at the end.

Deal with environmental variable: If the CORDOVA_BBTOOLS variable it
exists in the current terminal window is not going to affect it.
If it doesn't exists, it will be determinated and then settled, and at
the end of operations it will erase the value.
The meaning of this is, that it won't affect any other process or
execution in the terminal.