CB-5093: Add versionCode and CFBundleVersion during prepare
The version can be given as part of the <widget> tag.
Example:
<widget
	id="io.cordova.hellocordova" version="0.0.1"
	android-versionCode="7"
	ios-CFBundleVersion="3.3.3"
	...

If not specified in the widget tag, the following defaults will be used.
Assuming version = "MAJOR.MINOR.PATCH-whatever"
versionCode = PATCH + MINOR * 100 + MAJOR * 10000
CFBundleVersion = "MAJOR.MINOR.PATCH"

Not adding the "other" version for windows platforms at this stage.
It's is in
Properties/AssemblyInfo.cs which we currently don't modify.
[assembly: AssemblyVersion("1.0.0.0")]

For more info about the "other" versions see:
http://developer.android.com/tools/publishing/versioning.html
http://stackoverflow.com/questions/4933093/cfbundleversion-in-the-info-plist-upload-error