Windows 8.1 update breaks app manifest parser
The powers-that-be at Microsoft decided to add another namespace to the package.appxmanifest when I updated my Windows 8 app to Windows 8.1. The Package/Applications/Application/VisualElements element now has a namespace prefix of "m2:" which causes an error during project preparation. This commit contains a quick, low-risk update to accommodate that change.