Expose cordova-lib from cordova-cli
Downstream packages that depend on both cordova-lib and cordova-cli have two
copies of cordova-lib. One as direct dependency in node-modules/cordova-lib
and the other as cordova-cli's dep in
node_modules/cordova/node_modules/corodva-lib.
The two copies might end up being of different versions, but even if they are
identical it still results in two separate instances of the EventEmitter and
breaks event registration and handling in some cases.