Proper error code and message when failing (#295)
Prior to this, the CLI would exit with UnhandledPromiseRejectionWarning
and a full stack trace when an error occurred.

With this change, set an error code indicating failure and only display
the error message by default. The stack trace is printed with --verbose.

Furthermore, the loud-rejection handler is registered to fail properly
if any rejected promises snuck past our handling.