fix: cannot call config or create commands (#495)
A few exceptions aside, the `cli` function checks if the current command
exists as a property in `cordova-lib`. If not, it throws an error that
the command is unsupported:

    Cordova does not know config;
    try `cordova help` for a list of all the available commands.

Since the `config` and `create` commands are not handled by
`cordova-lib` anymore and their previous implementations in
`cordova-lib` have recently been removed, we need to make an exceptions
for those commands.