(android) Replace deprecated `android` command with `avdmanager`
The `android` command is deprecated. When you run the command, it displays the following:

```
*************************************************************************
The "android" command is deprecated.
For manual SDK, AVD, and project management, please use Android Studio.
For command-line tools, use tools/bin/sdkmanager and tools/bin/avdmanager
*************************************************************************
Running /path/to/android-sdk/tools/bin/avdmanager list target
```

On many systems the `android` command isn't even in the PATH and so attempting to run the original command fails.